Redirect multiple RTMP URLs
Some websites will present RTMP URLs from more than one host. It can be
difficult to determine which host is the correct one. Patch simply identifies
all unique hosts for the site and runs them all through the hosts file.

Additionally this site was able to invoke RtmpDump directly from RtmpSrv instead
of the normal separate invokation. This is a problem because the RtmpDump
spawned by RtmpSrv, while successful in downloading is not responsive to
keyboard commands such as Ctrl+C. RtmpDump is now killed if invoked by RtmpSrv,
then invoked manually as before.

Tested on
http://nickelodeon.nl/videos/36578-aflevering-1