Add some checks for RtmpSrv
Added a warning for proper version of RtmpSrv. Also fix this error

  rtmpsrv: command not found

This error occurs because many people put RtmpSrv in the same folder as the
script, which may or may not be on the PATH. RtmpSrv should be kept at

  /usr/local/bin

If the script finds RtmpSrv in the same folder, it will move it to the proper
location.