Allow any "jq" and "ffmpeg" on the PATH
Before the script was only allowing "jq" and "ffmpeg" found in /usr/local/bin.
This was the case because I did not want to involve /dev/null in the detection
of these programs.

I wrote a short function that searchs the output of a command for a given
string, in this case "command not found". This will allow for use of "jq" and
"ffmpeg" anywhere on the PATH. Another benefit is you can use the function to
search for Carriage Return "\r", to detect if a program is Windows Native or
Cygwin.