Monitor existence of folder rather than process memory
Previously the script monitored the memory usage of
Firefox (/proc/$arg_pid/maps). When the memory usage "flattened out" it was
taken as an indicator that the video had been loaded.

This method has proven to be unreliable in its current form across countries
and operating systems. This has been the cause of the much ballyhooed
"dumped too soon, retry" errors.

I noticed that while loading the video several files and folders are loaded in
the Firefox folder. The script now waits for the creation of "thumbnails"
folder before dumping Firefox. I ran 10 successful tests with this, but we shall
see how it goes.